feat: 更新用户表结构，支持BIGSERIAL主键和字段约束
- 将用户ID类型从uint改为int64，对应BIGSERIAL
- 添加数据库字段约束（邮箱格式、钱包地址长度、角色枚举等）
- 移除软删除字段，简化用户模型
- 更新所有相关服务、仓库、处理器和DTO
- 重新生成Swagger文档

+++ bondly-api/internal/dto/user.go
@@ -22,7 +22,7 @@ type UpdateUserRequest struct {
 
 // UserResponse 用户响应结构
 type UserResponse struct {
-	ID              uint    `json:"id" example:"1"`
+	ID              int64   `json:"id" example:"1"`
 	WalletAddress   *string `json:"wallet_address" example:"0x1234567890abcdef1234567890abcdef12345678"`
 	Email           *string `json:"email" example:"user@example.com"`
 	Nickname        string  `json:"nickname" example:"John Doe"`
